+++ src/main/java/com/example/umc8th/code/reply/service/command/ReplyCommandService.java
@@ -7,8 +7,6 @@ import com.example.umc8th.code.reply.entity.Reply;
 
 public interface ReplyCommandService {
     Reply createReply(ReplyRequestDTO.CreateReplyDTO dto, Long articleId);
-
-    Reply findActiveReply(Long articleId, Long replyId);
     Reply saveAndUpdate(Long articleId, ReplyRequestDTO.UpdateReplyDTO dto);
     void deleteReply(Long articleId, Long replyId);
 }
