+++ src/main/java/com/example/umc8th/code/reply/service/query/ReplyQueryService.java
@@ -8,4 +8,6 @@ public interface ReplyQueryService {
     List<Reply> getReplies();
     List<Reply> getRepliesByArticle(Long articleId);
 //    Reply getReply(Long id);
+    Reply findActiveReply(Long articleId, Long replyId);
+
 }
